Fix obvious problems with pyALS and pyAIS
Split into different files, fix crashes, fix spurious error values, and
add a getter/setter for ageInstanceGuid.
