+++ Python/CMakeLists.txt
@@ -323,7 +323,8 @@ set(PRP_MSG_HDRS
 )
 
 set(PRP_MISC_SRCS
-    PRP/Misc/pyAgeLinkInfo.cpp
+    PRP/Misc/pyAgeInfoStruct.cpp
+    PRP/Misc/pyAgeLinkStruct.cpp
     PRP/Misc/pyRenderLevel.cpp
 )
 set(PRP_MISC_HDRS
